+++ runtime/index.d.ts
@@ -40,10 +40,30 @@ declare module "@sapper/server" {
 	export function middleware(opts: MiddlewareOptions): Handler;
 }
 
-declare module "@sapper/service-worker" {
+declare module '@sapper/service-worker' {
 	export const timestamp: number;
 	export const files: string[];
 	export const assets: string[];
 	export const shell: string[];
 	export const routes: Array<{ pattern: RegExp }>;
+}
+
+declare module '@sapper/common' {
+	export interface PreloadContext {
+		fetch: (url: string, options?: any) => Promise<any>;
+		error: (statusCode: number, message: Error | string) => void;
+		redirect: (statusCode: number, location: string) => void;
+	}
+
+	export interface Page {
+		host: string;
+		path: string;
+		params: Record<string, string>;
+		query: Record<string, string | string[]>;
+		error?: Error;
+	}
+
+	export interface Preload {
+		(this: PreloadContext, page: Page, session: any): object | Promise<object>;
+	}
 }