+++ runtime/src/internal/manifest-client.d.ts
@@ -1,10 +1,10 @@
 import {
-  Preload
+	Preload
 } from './shared';
 
 export interface DOMComponentModule {
-  default: DOMComponentConstructor;
-  preload?: Preload;
+	default: DOMComponentConstructor;
+	preload?: Preload;
 }
 
 export interface DOMComponent {
@@ -13,19 +13,19 @@ export interface DOMComponent {
 }
 
 export interface DOMComponentConstructor {
-  new(options: { target: Element, props: unknown, hydrate: boolean }): DOMComponent;
+	new(options: { target: Element, props: unknown, hydrate: boolean }): DOMComponent;
 }
 
 export interface DOMComponentLoader {
-  js: () => Promise<DOMComponentModule>
+	js: () => Promise<DOMComponentModule>
 }
 
 export interface Route {
-  pattern: RegExp;
-  parts: Array<{
-    i: number;
-    params?: (match: RegExpExecArray) => Record<string, string>;
-  }>;
+	pattern: RegExp;
+	parts: Array<{
+		i: number;
+		params?: (match: RegExpExecArray) => Record<string, string>;
+	}>;
 }
 
 export const ErrorComponent: DOMComponentConstructor;