+++ site/src/utils/highlight.js
@@ -7,7 +7,7 @@ export function highlight(source, lang) {
 	const highlighted = plang ? PrismJS.highlight(
 		source,
 		PrismJS.languages[plang],
-		lang,
+		lang
 	) : source.replace(/[&<>]/g, c => ({ '&': '&amp;', '<': '&lt;', '>': '&gt;' })[c]);
 
 	return `<pre class='language-${plang}'><code>${highlighted}</code></pre>`;