+++ test/apps/scroll/src/routes/redirect.svelte
@@ -1,5 +1,5 @@
 <script context="module">
 	export function preload() {
-        this.redirect(301, 'another-tall-page')
-    }
+		this.redirect(301, 'another-tall-page');
+	}
 </script>