+++ test/apps/scroll/src/routes/search-form.svelte
@@ -1,12 +1,12 @@
 <script>
-	import { goto } from "@sapper/app";
+	import { goto } from '@sapper/app';
 
 	function preserveScroll() {
-		goto("/a-third-tall-page", { noscroll: true });
+		goto('/a-third-tall-page', { noscroll: true });
 	}
 
 	function scroll() {
-		goto("/a-third-tall-page");
+		goto('/a-third-tall-page');
 	}
 </script>